+++ app/Http/Controllers/ClanokController.php
@@ -3,21 +3,74 @@
 namespace App\Http\Controllers;
 
 use App\Article;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Input;
 use Illuminate\Support\Facades\Response;
-use Illuminate\Support\Facades\App;
 
 class ClanokController extends Controller
 {
-
-    public function getIndex()
+    public function getIndex(Request $request)
     {
-        $articles = Article::published()->orderBy('published_date', 'desc');
-        if (Input::has('author')) {
-            $articles = $articles->where('author', 'LIKE', Input::get('author'));
+        $articles = Article::published()->with('category');
+
+        // Filtering
+        if ($request->has('category')) {
+            $articles = $articles->whereHas('category', function (Builder $query) use ($request) {
+                $query->where('name', $request->input('category'));
+            });
         }
+
+        if ($request->has('author')) {
+            $articles = $articles->where('author', 'LIKE', $request->input('author'));
+        }
+
+        // Sorting
+        $sortBy = $request->input('sort_by', 'date_desc');
+
+        if ($sortBy === 'date_desc') {
+            $articles = $articles->orderBy('published_date', 'desc');
+        }
+
+        if ($sortBy === 'date_asc') {
+            $articles = $articles->orderBy('published_date', 'asc');
+        }
+
         $articles = $articles->get();
-        return view('clanky', array('articles'=>$articles));
+
+        // Options for filters
+        $categoriesOptions = $articles
+            ->countBy('category.name')
+            ->map(function ($count, $category) use ($request) {
+                return [
+                    'value' => $category,
+                    'text' => "$category ($count)",
+                    'selected' => $category === $request->input('category'),
+                ];
+            });
+
+        $authorsOptions = $articles
+            ->countBy('author')
+            ->map(function ($count, $author) use ($request) {
+                return [
+                    'value' => $author,
+                    'text' => "$author ($count)",
+                    'selected' => $author === $request->input('author'),
+                ];
+            });
+
+        $sortingOptions = collect([
+            [ 'value' => 'date_asc', 'text' => trans('articles.filter.sort_by.date_asc') ],
+            [ 'value' => 'date_desc', 'text' => trans('articles.filter.sort_by.date_desc') ],
+        ]);
+
+        return view('frontend.articles.index', compact(
+            'articles',
+            'categoriesOptions',
+            'authorsOptions',
+            'sortingOptions',
+            'sortBy'
+        ));
     }
 
     public function getSuggestions()
@@ -56,7 +109,7 @@ class ClanokController extends Controller
         $article->view_count += 1;
         $article->save();
 
-        return view('clanok', array('article'=>$article));
+        return view('frontend.articles.show', array('article'=>$article));
 
     }
 }